CXP-1165: move client from infra to settings
