+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/Settings/services.yml
@@ -1,0 +1,13 @@
+services:
+    Akeneo\Connectivity\Connection\Infrastructure\Settings\Client\Fos\CreateClient:
+        arguments:
+            - '@fos_oauth_server.client_manager.default'
+
+    Akeneo\Connectivity\Connection\Infrastructure\Settings\Client\Fos\DeleteClient:
+        arguments:
+            - '@fos_oauth_server.client_manager.default'
+
+    Akeneo\Connectivity\Connection\Infrastructure\Settings\Client\Fos\FosRegenerateClientSecret:
+        arguments:
+            - '@fos_oauth_server.client_manager.default'
+            - '@database_connection'
