+++ src/Akeneo/Connectivity/Connection/back/tests/Unit/spec/Infrastructure/Settings/Client/Fos/FosRegenerateClientSecretSpec.php
@@ -6,7 +6,7 @@ namespace spec\Akeneo\Connectivity\Connection\Infrastructure\Client\Fos;
 
 use Akeneo\Connectivity\Connection\Application\Settings\Service\RegenerateClientSecretInterface;
 use Akeneo\Connectivity\Connection\Domain\Settings\Model\ValueObject\ClientId;
-use Akeneo\Connectivity\Connection\Infrastructure\Client\Fos\FosRegenerateClientSecret;
+use Akeneo\Connectivity\Connection\Infrastructure\Settings\Client\Fos\FosRegenerateClientSecret;
 use Akeneo\Tool\Bundle\ApiBundle\Entity\Client;
 use Doctrine\DBAL\Connection as DbalConnection;
 use Doctrine\DBAL\Driver\Statement;
